pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "drape-backend"
version = "0.1.0"
requires-python = ">=3.12"
dependencies = [
    "fastapi==0.115.6",
    "uvicorn[standard]==0.34.0",
    "dramatiq[redis]==1.17.0",
    "supabase==2.28.3",
    "cloudinary==1.41.0",
    "google-generativeai==0.8.3",
    "modal==0.73.0",
    "Pillow==11.1.0",
    "python-multipart==0.0.20",
    "httpx==0.28.1",
    "python-jose[cryptography]==3.3.0",
    "pydantic==2.11.7",
    "pydantic-settings==2.8.1",
    "transformers==4.48.1",
    "ftfy==6.3.1",
    "open_clip_torch==2.29.0",
    "torch==2.5.1",
    "numpy==2.2.1",
]

[project.optional-dependencies]
dev = [
    "pytest==8.3.4",
    "pytest-asyncio==0.25.2",
    "httpx==0.28.1",
    "respx==0.22.0",
]

[tool.hatch.build.targets.wheel]
packages = ["app"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
